cleanup

+++ python/ray/tune/function_runner.py
@@ -189,8 +189,7 @@ class FunctionRunner(Trainable):
         session.init(self._status_reporter)
         self._runner = None
         self._restore_tmpdir = None
-        self.default_checkpoint_dir = TrainableUtil.make_checkpoint_dir(
-            self.logdir, index="default", dryrun=True)
+        self.default_checkpoint_dir = None
 
     def _trainable_func(self):
         """Subclasses can override this to set the trainable func."""
@@ -284,10 +283,8 @@ class FunctionRunner(Trainable):
         return fn(self)
 
     def create_default_checkpoint_dir(self):
-        checkpoint_dir = TrainableUtil.make_checkpoint_dir(
+        self.default_checkpoint_dir = TrainableUtil.make_checkpoint_dir(
             self.logdir, index="default")
-        assert checkpoint_dir == self.default_checkpoint_dir, (
-            "Checkpoint directory validation failed!")
         return self.default_checkpoint_dir
 
     def save(self, checkpoint_path=None):
@@ -322,7 +319,8 @@ class FunctionRunner(Trainable):
         self._status_reporter.save_checkpoint(checkpoint)
 
     def restore_from_object(self, obj):
-        if os.exists(self.default_checkpoint_dir):
+        if self.default_checkpoint_dir is not None and os.exists(
+                self.default_checkpoint_dir):
             shutil.rmtree(self.default_checkpoint_dir)
             logger.debug("Clearing default checkpoint: %s",
                          self.default_checkpoint_dir)
@@ -341,6 +339,12 @@ class FunctionRunner(Trainable):
         # Check for any errors that might have been missed.
         self._report_thread_runner_error()
         session.shutdown()
+
+        if self.default_checkpoint_dir is not None and os.exists(
+                self.default_checkpoint_dir):
+            shutil.rmtree(self.default_checkpoint_dir)
+            logger.debug("Clearing default checkpoint: %s",
+                         self.default_checkpoint_dir)
 
     def _report_thread_runner_error(self, block=False):
         try:
